requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data handling
yfinance>=0.2.0
pandas-ta>=0.3.14b
pandas-market-calendars>=4.0.0
pyarrow>=10.0.0

# API clients
requests>=2.28.0
aiohttp>=3.8.0
websockets>=9.0,<11.0
alpha-vantage>=2.3.1
tiingo>=0.14.0

# Database and caching
sqlalchemy>=2.0.0
redis>=4.0.0
pymongo>=4.0.0

# Execution and broker APIs
ccxt>=3.0.0  # Crypto exchanges
alpaca-trade-api>=3.0.0
ib-insync>=0.9.0  # Interactive Brokers

# Monitoring and visualization
plotly>=5.0.0
dash>=2.0.0
streamlit>=1.20.0
prometheus-client>=0.16.0

# Risk and portfolio management
pyfolio>=0.9.2
quantstats>=0.0.59
empyrical>=0.5.5

# Configuration and logging
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Task scheduling
apscheduler>=3.10.0
celery>=5.2.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Type checking and linting
mypy>=1.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
